ugc_etl/requirements.txt
clickhouse-driver==0.2.6
kafka-python==2.0.2
